+++ pyobjc-framework-AuthenticationServices/PyObjCTest/test_asauthorizationerror.py
@@ -1,0 +1,11 @@
+from PyObjCTools.TestSupport import *
+
+import AuthenticationServices
+
+class TestASAuthorization (TestCase):
+    def test_constants(self):
+        self.assertEqual(AuthenticationServices.ASAuthorizationErrorUnknown, 1000)
+        self.assertEqual(AuthenticationServices.ASAuthorizationErrorCanceled, 1001)
+        self.assertEqual(AuthenticationServices.ASAuthorizationErrorInvalidResponse, 1002)
+        self.assertEqual(AuthenticationServices.ASAuthorizationErrorNotHandled, 1003)
+        self.assertEqual(AuthenticationServices.ASAuthorizationErrorFailed, 1004)