+++ pyobjc-framework-AuthenticationServices/PyObjCTest/test_asauthorizationprovider.py
@@ -1,0 +1,12 @@
+from PyObjCTools.TestSupport import *
+
+import AuthenticationServices
+
+class TestASAuthorizationProvider (TestCase):
+    @min_sdk_version('10.15')
+    def test_protocols(self):
+        objc.protocolNamed('ASAuthorizationProvider')
+
+    @min_os_level('10.15')
+    def test_methods10_15(self):
+        self.assertArgIsBlock(AuthenticationServices.ASAuthorizationProviderExtensionAuthorizationRequest.presentAuthorizationViewControllerWithCompletion_, 0, b'vZ@')