+++ pyobjc-framework-CoreML/Modules/_CoreML_protocols.m
@@ -9,4 +9,7 @@ static void __attribute__((__used__)) use_protocols(void)
     p = PyObjC_IdToPython(@protocol(MLCustomModel)); Py_XDECREF(p);
     p = PyObjC_IdToPython(@protocol(MLBatchProvider)); Py_XDECREF(p);
 #endif
+#if PyObjC_BUILD_RELEASE >= 1015 && defined(__x86_64__)
+    p = PyObjC_IdToPython(@protocol(MLWritable)); Py_XDECREF(p);
+#endif
 }